finished project structure
