+++ src/resources/api/v1/business/business.controller.js
@@ -1,6 +1,6 @@
-import BusinessService from "../../../../services/business.service"
+import BusinessService from "./business.service"
 const {createBusiness} = BusinessService
-import BusCategoryService from "../../../../services/businessCategory.service"
+import BusCategoryService from "../businesscategory/businessCategory.controller"
 import customMessage from "../../../../utils/customMessage";
 import statusCode from "../../../../utils/statusCode";
 import responses from "../../../../utils/responses";
