+++ src/resources/api/v1/business/business.service.js
@@ -1,4 +1,4 @@
-import models from "../database/models/index"
+import models from "../../../../database/models/index"
 const{Busines}=models
 
 export default class BusinessCategory{
