+++ src/resources/api/v1/index.js
@@ -1,15 +1,11 @@
 import express from "express";
-import Business from "./business/business.route"
-import BusinessCategory from "./business/businessCategory.routes"
-import Owner from "./owner/owner.route"
+import Business from "./business/business.routes"
+import BusinessCategory from "./businesscategory/businessCategory.routes"
+import Owner from "./owner/owner.routes"
 
 const router = express.Router();
 router.use("/owner", Owner)
-router.use("/user", User);
 router.use("/business_category", BusinessCategory)
 router.use("/business", Business)
-router.use('/rolesPermissions', rolePerm);
-router.use('/roles', roleRoutes);
-router.use('/permissions', permissionRoutes);
 
 export default router;
