+++ src/resources/api/v1/owner/businessOwner.service.js
@@ -1,4 +1,4 @@
-import models from "../database/models/index.js";
+import models from "../../../../database/models/index.js";
 import "regenerator-runtime/runtime";
 import validator from "validator"
