--- test/data/user.mock.js
@@ -1,64 +1,0 @@
-export default {
-  user1: {
-    firstname: "Duncan",
-    lastname: "warrior",
-    email: "klf89416@cuoly.com",
-    password: "Password1!",
-    username: "warrior2"
-  },
-  userTest:{
-    firstname: "Duncan",
-    lastname: "warrior",
-    email: "klf89416",
-    password: "Password1!",
-    username: "warrior444",
-  },
-  password1: { password: "tytyne1234", confirmPassword: "tytyne12345" },
-  password2: { password: "123", confirmPassword: "123" },
-  password3: { password: "tytyne1234", confirmPassword: "tytyne1234" },
-
-  user2: {
-    firstname: "Duncan",
-    lastname: "warrior",
-    email: "klf89416@cdd.com",
-    password: "Password1!",
-    username: "warrior4",
-  },
-  user3: {
-    id: 1,
-    email: "user1@example.com"
-  },
-  user4: {
-    id: 2,
-    email: "user2@example.com"
-  },
-  user0: {
-    id: 0,
-    email: "user0@example.com"
-  },
-
-  password: {
-    password: "tytyne12345",
-    confirmPassword: "tytyne12345"
-  },
-  user6: { email: "lele@gmail.com" },
-  password1: { password: "tytyne1234", confirmPassword: "tytyne12345" },
-  password2: { password: "123", confirmPassword: "123" },
-  password3: { password: "tytyne1234", confirmPassword: "tytyne1234" },
-  user1Login: {
-    email: "klf89416@cdd.com",
-    password: "Password1!",
-  },
-  user1LoginWrongPassword: {
-    email: "klf89416@cdd.com",
-    password: "Password1",
-  },
-
-  user7: {
-    firstname: "Duncan",
-    lastname: "warrior",
-    email: "klf89416",
-    password: "Password1!",
-    username: "warrior77",
-  },
-};
